added execution script for use via jenkins refactored some generally useful functions into /application/deed/utils
