+++ schema_test.sh
@@ -1,0 +1,1 @@
+python3 schema_tests.py
