+++ schema_tests/utils.py
@@ -1,24 +1,5 @@
-from underscore import _
 import json
-import sys
 import re
-
-
-def down_one_level(schema, key, context):
-    try:
-        res = schema[key]
-        return res
-    except:
-        print("ACCESS ERROR:\nlocation in schema: %s\n with key: %s \n%s."
-              % (schema, key, sys.exc_info()[0]))
-        raise
-
-
-# schema is a json obj/dict
-# path - contains the string path to a dict attribute
-# with '/' separators e.g. /root/sub_ele/child1/attr_name
-def get_obj_by_path(schema, path):
-    return _.reduce(path.strip("/").split("/"), down_one_level, schema)
 
 
 def load_json_file(file_name):
